+++ analysis/analysers/parsers/UnifiedParser/src/main/kotlin/de/maibornwolff/codecharta/analysers/parsers/unified/metriccollectors/TreeSitterAdapter.kt
@@ -57,6 +57,7 @@ object TreeSitterAdapter {
         FileExtension.GO -> Language.GO
         FileExtension.PHP -> Language.PHP
         FileExtension.RUBY -> Language.RUBY
+        FileExtension.SCALA -> Language.SCALA
         FileExtension.SWIFT -> Language.SWIFT
         FileExtension.BASH -> Language.BASH
         FileExtension.CSHARP -> Language.CSHARP