+++ analysis/analysers/parsers/UnifiedParser/src/test/kotlin/de/maibornwolff/codecharta/analysers/parsers/unified/UnifiedParserTest.kt
@@ -43,6 +43,8 @@ class UnifiedParserTest {
             Arguments.of("php", ".php"),
             Arguments.of("python", ".py"),
             Arguments.of("ruby", ".rb"),
+            Arguments.of("scala", ".scala"),
+            Arguments.of("sbt", ".sbt"),
             Arguments.of("swift", ".swift"),
             Arguments.of("typescript", ".ts"),
             Arguments.of("tsx", ".tsx"),